Further Implement alternative Delaunay Triangulation

+++ Delaunay/Geometry.pde
@@ -7,7 +7,7 @@ double angleBetween(Point p1, Point p2, Point p3)
   //Second vector
   Point v2 = new Point((p3.x-p2.x),(p3.y-p2.y));
   
-  int dP = v1.x *v2.x +v1.y*v2.y;
+  double dP = v1.x *v2.x +v1.y*v2.y;
   double v1Norm= Math.sqrt(v1.x*v1.y +v1.x*v1.x);
   double v2Norm = Math.sqrt(v2.x*v2.y +v2.x*v2.x);
